% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omotomo_data.R
\name{omotomo_data}
\alias{omotomo_data}
\title{Read and process data for tomography scan}
\usage{
omotomo_data(t0, t1, tomo = NULL, verbose = FALSE)
}
\arguments{
\item{t0}{\link{POSIXct}
Start time of the analysis interval.}

\item{t1}{\link{POSIXct}
End time of the analysis interval.}

\item{verbose}{\link{logical} (default: \code{FALSE})
If \code{TRUE}, prints progress messages and diagnostic information.}
}
\value{
Updates and returns the global \code{tomo} list, containing fields such as:
\itemize{
\item \verb{tomo$obs_ne_peaks - Ionospheric peak parameters \\item }tomo$obs_stec - Slant TEC measurement (ground-based GNSS, Radio occultation)
\item `tomo$obs_vtec - Vertica TEC measurements (Altimeter Jason/Topex)
}
}
\description{
Read and process data for tomography scan
}
\details{
The function performs the following main steps:
\enumerate{
\item Collects and processes multi-instrument observational data:
\itemize{
\item GNSS (Madrigal STEC data)
\item Radio occultation (STEC and profiles)
\item Ionosonde profiles
\item Incoherent scatter radar (ISR) profiles
\item DIDBase parameters
}
\item Extracts ionospheric peak parameters and smooths time series.
}

Errors are thrown with class \code{"tomoscand_error"} if required inputs or files
are missing, invalid, or incorrectly specified.
}
