% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omotomo.R
\name{parse_omotomo_filename_time}
\alias{parse_omotomo_filename_time}
\title{Parse Omotomo file end time from filename}
\usage{
parse_omotomo_filename_time(fname, tz = "UTC")
}
\arguments{
\item{fname}{Character. Filename or full path.}

\item{tz}{Character. Timezone used for output, default "UTC".}
}
\value{
POSIXct. Parsed file end time.
}
\description{
Omotomo filenames are expected to follow the pattern
\verb{omotomo<yydoyHHMM>_par.h5}, where:
\itemize{
\item \code{yy} is the 2 digit year (assumed 2000 + yy)
\item \code{doy} is day of year (001 to 366)
\item \code{HHMM} is hour and minute in UTC
}
}
\details{
The parsed time corresponds to the last sample time in the file.
}
